+++ src/app/tasks/interfaces/index.ts
@@ -5,3 +5,5 @@ export * from './new-task-request.interface';
 export * from './delete-response.interface';
 export * from './mini-task-btn.type';
 export * from './mini-task-action-response.interface';
+export * from './update-task-request.interface';
+export * from './form-detail.interface';