+++ src/app/tasks/pages/task-details/task-details.component.ts
@@ -12,15 +12,22 @@ import { rxResource } from '@angular/core/rxjs-interop';
 
 import { map } from 'rxjs';
 import { TranslateModule, TranslateService } from '@ngx-translate/core';
+import { LucideAngularModule } from 'lucide-angular';
+import moment from 'moment';
 
 import { FormatDateService, HeaderService, TaskService } from '../../services';
 import { MessageService } from '../../../shared/services';
-import { TaskResponseItem } from '../../interfaces';
+import { TaskResponseItem, TaskState } from '../../interfaces';
 import { TaskLogSectionComponent } from '../../components/task-log-section/task-log-section.component';
 
 @Component({
   selector: 'app-task-details',
-  imports: [TranslateModule, ReactiveFormsModule, TaskLogSectionComponent],
+  imports: [
+    TranslateModule,
+    ReactiveFormsModule,
+    LucideAngularModule,
+    TaskLogSectionComponent,
+  ],
   templateUrl: './task-details.component.html',
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
@@ -46,6 +53,7 @@ export default class TaskDetailsComponent {
   setFormDataEffect = effect(() => {
     const taskDetailsData = this.taskDetailValue();
     if (taskDetailsData) {
+      this.setTitle(taskDetailsData.name);
       this.setFormValue(taskDetailsData);
     }
   });
@@ -53,7 +61,7 @@ export default class TaskDetailsComponent {
   taskDetailsForm = this.fb.group({
     name: [''],
     description: [''],
-    currentState: [''],
+    currentState: [TaskState.CREATED],
     currentStateTranslated: [''],
     startDate: [''],
     startTime: [''],
@@ -62,9 +70,19 @@ export default class TaskDetailsComponent {
   });
 
   constructor() {
-    const headerTitle = this.translate.instant('taskDetails.title');
-    this.headerService.setTitle(headerTitle);
+    this.setHeaderTitle();
+  }
+
+  private setHeaderTitle(): void {
     this.headerService.showBackBtn = true;
+    this.setTitle('');
+  }
+
+  private setTitle(taskName: string): void {
+    const headerTitle = this.translate.instant('taskDetails.title', {
+      taskName,
+    });
+    this.headerService.setTitle(headerTitle);
   }
 
   setFormValue(taskDetailData: TaskResponseItem): void {
@@ -79,13 +97,13 @@ export default class TaskDetailsComponent {
         `taskStates.${currentState}`
       ),
       startDate: startDate
-        ? this.formatDateService.getDateFromDate(startDate)
+        ? this.formatDateService.getDate(startDate, false)
         : '',
       startTime: startDate
-        ? this.formatDateService.getTimeFromDate(startDate)
+        ? this.formatDateService.getTime(startDate, false)
         : '',
-      dueDate: dueDate ? this.formatDateService.getDateFromDate(dueDate) : '',
-      dueTime: dueDate ? this.formatDateService.getTimeFromDate(dueDate) : '',
+      dueDate: dueDate ? this.formatDateService.getDate(dueDate, false) : '',
+      dueTime: dueDate ? this.formatDateService.getTime(dueDate, false) : '',
     });
   }