+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.9.0] - 2025-07-18
+
+- Added update task page.
+- Change styles for non midifiable form fields.
+- Added some more icons.
+- Added close button in update task.
+- Added name in task detail and task update pages.
+- Change new task styles in form.
+- Bug fixed.
+
 ## [0.8.0] - 2025-07-13
 
 - Added task details page.